Fix compiler warning: implicit conversion loses integer precision
src/gsb_real.c:891:17: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
    nbre_char = strlen ( str_number );
              ~ ^~~~~~~~~~~~~~~~~~~~~
src/gsb_real.c:895:23: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
    if ( ( longueur = strlen ( thousands_sep ) ) == 1 )
                    ~ ^~~~~~~~~~~~~~~~~~~~~~~~

+++ src/gsb_real.c
@@ -879,11 +879,11 @@ gchar *gsb_real_add_thousands_sep ( gchar *str_number, const gchar *thousands_se
     gchar *dest;
     gchar *tmp_ptr;
     gchar **tab_str = NULL;
-    gint nbre_char;
+    gssize nbre_char;
     gint i = 0;
     gint j = 0;
     gint sep = 0;
-    gint longueur;
+    gssize longueur;
 
     if ( thousands_sep == NULL )
         return str_number;
